Refactor; set up strict schema validation
