+++ middleware/schema_and_dto/dynamic/schema/doc_construction/mappings.py
@@ -3,5 +3,4 @@ from utilities.enums import SourceMappingEnum
 PARSER_SOURCE_LOCATION_MAP = {
     SourceMappingEnum.QUERY_ARGS: "query",
     SourceMappingEnum.PATH: "path",
-    SourceMappingEnum.FILE: "file",
 }
