+++ middleware/schema_and_dto/dynamic/schema/request_content_population_/source_extraction/core.py
@@ -1,20 +1,12 @@
-from typing import Callable
-
-from flask import request
-
-
+from middleware.schema_and_dto.dynamic.schema.request_content_population_.source_extraction.mapping import \
+    EXTRACTOR_MAPPING
+from middleware.schema_and_dto.dynamic.schema.request_content_population_.types import JSONDict
 from utilities.enums import SourceMappingEnum
 
 
-def _get_source_getting_function(source: SourceMappingEnum) -> Callable:
-    source_function_mapping: dict[SourceMappingEnum, Callable] = {
-        SourceMappingEnum.QUERY_ARGS: request.args.get,
-        SourceMappingEnum.FORM: request.form.get,
-        SourceMappingEnum.JSON: lambda key: (
-            request.json.get(key) if request.json else None
-        ),
-        SourceMappingEnum.PATH: request.view_args.get,  # pyright: ignore[reportOptionalMemberAccess]
-        SourceMappingEnum.FILE: request.files.get,
-    }
-
-    return source_function_mapping[source]
+def get_data_from_source(source: SourceMappingEnum, fields: list[str]) -> JSONDict:
+    extractor_class = EXTRACTOR_MAPPING[source]
+    extractor = extractor_class(
+        expected_fields=fields,
+    )
+    return extractor.data
