+++ middleware/util/argument_checking.py
@@ -1,8 +1,7 @@
-from typing import Optional, Any
+from typing import Any
 
 from utilities.argument_checking_logic import (
     MutuallyExclusiveArgumentError,
-    MissingRequiredArgumentError,
 )
