+++ tests/middleware/request_content_population/populate_with_request_content/dto/test_no_transformation_function.py
@@ -1,0 +1,15 @@
+from middleware.schema_and_dto.dynamic.dto_request_content_population import populate_dto_with_request_content
+from middleware.schema_and_dto.non_dto_dataclasses import DTOPopulateParameters
+from tests.middleware.request_content_population.data import SimpleDTO
+from utilities.enums import SourceMappingEnum
+
+
+def test_populate_dto_with_request_no_transformation_functions(
+    patched_get_data_from_source,
+):
+    dto = populate_dto_with_request_content(
+        DTOPopulateParameters(dto_class=SimpleDTO, source=SourceMappingEnum.QUERY_ARGS)
+    )
+    assert dto.simple_string == "spam"
+    assert dto.optional_int is None
+    assert dto.transformed_array == "hello,world"
